Major update for character-level input preparation and CharCNN:
preprocessor.py
- "is_char_level" parameter for Preprocessor constructor. If it is true,
preprocessing will be done according to characters. Otherwise, sentence/word level
preprocessing will be done.
- Two new methods for character-level preprocessing. One is for defining
"<SPACE>" (whitespace) as a special character and the other is for defining
"<NUM>" (for each digit character) as a special character.
- preprocess method is extended to handle new methods.

dataset_reader.py
- Character-level data reading is not one of the better sides of torchtext,
hence a minor, evil hack method is added: "evil_workaround_tokenizer". The
reasoning of the method is in code as comment =)
- sentence_field can be created as a list of chars, and special tokens, of a sentence,
now (in word-level it is list of words and special tokens)
- Main method for test purposes inside this code is changed, too.

CNN.py
- CharCNN's forward method now handle input's shape correctly.
- Kl_loss object has a .device() now

main.py
- Due to changes in Preprocessor objects, main flow is changed, too.
- Probably some of the hard-coded paths changed again but I think you don't care =)